fix: move removing file to readwriter
